do not remove ProductPropertyValues that belong to multiple groups attached to single products for multiple times

+++ lfs/catalog/listeners.py
@@ -44,7 +44,8 @@ def property_group_deleted_listener(sender, instance, **kwargs):
 
     for product in products:
         for prop in properties:
-            ProductPropertyValue.objects.filter(product=product, property=prop).delete()
+            if prop.groups.count() == 1:
+                ProductPropertyValue.objects.filter(product=product, property=prop).delete()
 pre_delete.connect(property_group_deleted_listener, sender=PropertyGroup)
